Move port_allocator to apimachinery
